requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
billiard==3.5.0.5
celery==4.2.2
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cloudinary==1.15.0
coverage==4.5.2
coveralls==1.6.0
cryptography==2.6.1
Django==2.1.7
django-appconf==1.0.3
django-celery-email==2.0.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.8
kombu==4.3.0
mock==2.0.0
pbr==5.1.3
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.3.0
